# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Residency Scheduler API
  description: |
    Comprehensive API for medical residency program scheduling with ACGME compliance,
    emergency coverage management, and advanced resilience features.

    ## Authentication
    All endpoints require JWT Bearer token authentication except `/auth/login` and `/auth/register`.

    Access tokens are stored in httpOnly cookies and can also be provided via Authorization header:
    ```
    Authorization: Bearer <ACCESS_TOKEN>
    ```

    ## Base URL
    ```
    http://localhost:8000/api
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: Apache 2.0
servers:
  - url: http://localhost:8000/api
    description: Development server
  - url: https://api.example.com/api
    description: Production server
paths:
...
  /assignments:
    get:
      tags:
        - Assignments
      summary: List assignments
      security:
        - bearerAuth: []
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: person_id
          in: query
          schema:
            type: string
            format: uuid
        - name: role
          in: query
          schema:
            type: string
        - name: rotation_type
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: List of assignments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentListResponse'

    post:
      tags:
        - Assignments
      summary: Create assignment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentCreate'
      responses:
        '201':
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentWithWarnings'
        '422':
          description: ACGME violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentWithWarnings'

    delete:
      tags:
        - Assignments
      summary: Delete assignments in date range
      security:
        - bearerAuth: []
      parameters:
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '204':
          description: Assignments deleted
  /assignments/{assignment_id}:
    get:
      tags:
        - Assignments
      summary: Get assignment
      security:
        - bearerAuth: []
      parameters:
        - name: assignment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Assignment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentResponse'

    put:
      tags:
        - Assignments
      summary: Update assignment
      security:
        - bearerAuth: []
      parameters:
        - name: assignment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentUpdate'
      responses:
        '200':
          description: Assignment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentWithWarnings'

    delete:
      tags:
        - Assignments
      summary: Delete assignment
      security:
        - bearerAuth: []
      parameters:
        - name: assignment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Assignment deleted

  # ============================================================================
  # Schedule Endpoints
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AssignmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        person_id:
          type: string
          format: uuid
        person_name:
          type: string
        block_id:
          type: string
          format: uuid
        block_date:
          type: string
          format: date
        session:
          type: string
        rotation_template_id:
          type: string
          format: uuid
        rotation_name:
          type: string
        role:
          type: string
        activity_name:
          type: string
        status:
          type: string
    AssignmentCreate:
      type: object
      properties:
        person_id:
          type: string
          format: uuid
        block_id:
          type: string
          format: uuid
        rotation_template_id:
          type: string
          format: uuid
        role:
          type: string
        override_reason:
          type: string
      required:
        - person_id
        - block_id
        - rotation_template_id
        - role
    AssignmentUpdate:
      type: object
      properties:
        person_id:
          type: string
          format: uuid
        block_id:
          type: string
          format: uuid
        rotation_template_id:
          type: string
          format: uuid
        role:
          type: string
        version:
          type: integer
    AssignmentWithWarnings:
      type: object
      properties:
        assignment:
          $ref: '#/components/schemas/AssignmentResponse'
        warnings:
          type: array
          items:
            type: object
        violations:
          type: array
          items:
            type: object
    AssignmentListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AssignmentResponse'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer

    # Schedule Schemas
...